Split run_demo into try-along step scripts and add a setup guide
Break the monolithic run_demo.py pipeline into three standalone, runnable
step scripts that mirror the README's three phases, so people can run one
phase at a time and follow along:

  step1_intake.py      referral note -> FHIR bundle -> IPS -> referral agent -> EHR write
  step2_evaluate.py    BCBS-MA policy #297 agent evaluates the PA (+ the evidence flip)
  step3_adjudicate.py  extract CPT codes, adjudicate APPROVED / DENIED

- common.py: shared config/auth/helpers (moved out of run_demo.py) plus a
  cleanup() helper, a gitignored .state/ artifact store the steps pass data
  through, and a require() guard for out-of-order runs.
- run_demo.py is now a thin orchestrator that chains the three step run()
  functions with one shared state and a single cleanup (no logic duplication).
- evals/run_evals.py imports helpers from common instead of run_demo.
- Add requirements.txt (phenoml, python-dotenv, httpx); replace the dead
  .cache_bundle.json with .state/ in .gitignore.
- README: ordered getting-started (venv, requirements, provider discovery)
  plus a "Run the demo" try-along section.

Also adds the Part 1 / Part 2 course modules the README links to.

Verified end-to-end against a live instance: step1->step2->step3 with state
passed via .state/, the decision flip (true->false), APPROVED/DENIED
adjudication, the out-of-order guard, and per-step agent cleanup.
